LayoutTests/imported/w3c/web-platform-tests/worklets/animation-worklet-service-worker-interception.https-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'get_worklet(worklet_type).addModule')
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'get_worklet(worklet_type).addModule')
